fix: updates dependencies and documentation
Updates various dependencies in `package.json` files for the `docs`, `nova-core`, and `nova-transactions` packages to their latest versions.

Refactors the `NovaProvider` documentation to provide a more comprehensive usage guide, including code examples and explanations of how to integrate it with a Pulsar store and wallet connector.

Adds documentation for `nova-transactions` README.md